Move frequency_band_helpers to processing
- update tests imports
